chore: modularize data models
