+++ bc_obps/registration/models/__init__.py
@@ -1,0 +1,56 @@
+# Order of imports is important to avoid circular dependencies
+from .time_stamped_model import TimeStampedModel
+from .user_and_contact_common_info import UserAndContactCommonInfo
+from .address import Address
+from .app_role import AppRole
+from .bc_obps_regulated_operation import BcObpsRegulatedOperation
+from .business_role import BusinessRole
+from .business_structure import BusinessStructure
+from .document_type import DocumentType
+from .document import Document
+from .contact import Contact
+from .user import User
+from .operator import Operator
+from .user_operator import UserOperator
+from .regulated_product import RegulatedProduct
+from .naics_code import NaicsCode
+from .reporting_activity import ReportingActivity
+from .operation_type import OperationType
+from .operation import Operation
+from .facility_type import FacilityType
+from .facility import Facility
+from .event import Event
+from .facility_ownership_timeline import FacilityOwnershipTimeline
+from .multiple_operator import MultipleOperator
+from .operation_ownership_timeline import OperationOwnershipTimeline
+from .parent_operator import ParentOperator
+from .well_authorization_number import WellAuthorizationNumber
+
+__all__ = [
+    "Address",
+    "AppRole",
+    "BcObpsRegulatedOperation",
+    "BusinessRole",
+    "BusinessStructure",
+    "Contact",
+    "DocumentType",
+    "Document",
+    "Event",
+    "FacilityOwnershipTimeline",
+    "FacilityType",
+    "Facility",
+    "MultipleOperator",
+    "NaicsCode",
+    "OperationOwnershipTimeline",
+    "OperationType",
+    "Operation",
+    "Operator",
+    "ParentOperator",
+    "RegulatedProduct",
+    "ReportingActivity",
+    "TimeStampedModel",
+    "UserAndContactCommonInfo",
+    "UserOperator",
+    "User",
+    "WellAuthorizationNumber",
+]
