+++ bc_obps/registration/models/address.py
@@ -1,0 +1,31 @@
+from common.models import BaseModel
+from django.db import models
+from localflavor.ca.models import CAPostalCodeField, CAProvinceField
+from simple_history.models import HistoricalRecords
+
+
+class Address(BaseModel):
+    """Address model"""
+
+    street_address = models.CharField(
+        max_length=1000, null=True, blank=True, db_comment="Street address of relevant location)"
+    )
+    municipality = models.CharField(
+        max_length=1000, null=True, blank=True, db_comment="Municipality of relevant location"
+    )
+    province = CAProvinceField(
+        db_comment="Province of the relevant location, restricted to two-letter province postal abbreviations",
+        null=True,
+        blank=True,
+    )
+    postal_code = CAPostalCodeField(
+        db_comment="Postal code of relevant location, limited to valid Canadian postal codes", null=True, blank=True
+    )
+    history = HistoricalRecords(
+        table_name='erc_history"."address_history',
+        history_user_id_field=models.UUIDField(null=True, blank=True),
+    )
+
+    class Meta:
+        db_table_comment = "Table containing address data. Only Canadian addresses are supported."
+        db_table = 'erc"."address'
